requirements.txt
qrcode[pil]
opencv-python
Pillow
requests
